requirements.txt
Flask
Flask-CORS
python-dotenv
google-play-scraper
textblob
transformers 
torch 
openpyxl
pandas
openai